split off URI into its own module
