+++ Sources/Biome/Ecosystem/Ecosystem.Resolution.swift
@@ -1,6 +1,7 @@
 import Versions
 import Resources
 import DOM
+import URI
 
 extension Ecosystem 
 {
